Make libcanberra sound theme configurable
On Linux, kitty uses libcanberra to play an audible bell following the
XDG Sound Theming Specification. The library expects the application to
include the sound theme in the context, otherwise it plays from the
"freedesktop" theme; it does not query the environment or configuration.

Rather than link against Gtk and query GtkSettings, this change allows
the sound theme to be configured with a new kitty setting,
"linux_bell_theme_name".